dcp-t5(002): move cli/workflow under factory_runtime
Relocate pkg/transports/cli/workflow to
pkg/services/factory_runtime/transports/cli/workflow. The package imports
only factory_runtime among product services, so it belongs in that
service's own transports subtree.

Pure relocation: only the two self-import paths in preview_test.go and
validate_test.go change. The package has no external Go importers. The
documented verification command in docs/reference/sessions.md is
retargeted so it still runs, and the unit coverage minimum moves with the
package.
